@@ -1,18 +1,29 @@
 apply plugin: 'com.android.application'
+apply plugin: 'kotlin-kapt'
+apply plugin: 'kotlin-android'
+apply plugin: 'kotlin-android-extensions'
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion "23.0.3"
+    compileSdkVersion 27
+    buildToolsVersion "27.0.3"
 
     defaultConfig {
         applicationId "com.example.abhim.popularmovies"
         minSdkVersion 17
-        targetSdkVersion 23
+        targetSdkVersion 27
         versionCode 1
         versionName "1.0"
         manifestPlaceholders = [HOCKEYAPP_APP_ID: "0f781253e2fa404789a78b55ccbd0504"]
 
+        jackOptions {
+            enabled true
+        }
     }
+
+    kapt {
+        generateStubs = true
+    }
+
     buildTypes {
         release {
             minifyEnabled false
@@ -21,6 +32,21 @@ android {
     }
     buildTypes.each {
         it.buildConfigField 'String', 'POPULAR_MOVIES_API_KEY', "\"1f19c0f420eab0495d79f856d09c9a29\""
+        it.buildConfigField 'String', 'BASE_URL', '"https://api.themoviedb.org/3/"'
+    }
+
+    sourceSets {
+        main.java.srcDirs += 'src/main/kotlin'
+    }
+
+    packagingOptions {
+        exclude 'META-INF/rxjava.properties'
+    }
+
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+        incremental false
     }
 }
 repositories {
@@ -30,12 +56,29 @@ repositories {
 dependencies {
     compile fileTree(include: ['*.jar'], dir: 'libs')
     testCompile 'junit:junit:4.12'
-    compile 'com.android.support:appcompat-v7:23.4.0'
+
+    //kotlin
+    compile "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
+    compile 'com.android.support:support-annotations:27.1.1'
+    compile 'com.android.support:appcompat-v7:27.1.1'
     compile 'com.squareup.picasso:picasso:2.5.2'
-    compile 'com.android.support:support-v4:23.4.0'
+    compile 'com.android.support:support-v4:27.1.1'
     compile 'net.hockeyapp.android:HockeySDK:4.1.0'
     compile 'com.jakewharton:butterknife:5.1.1'
-    compile 'com.squareup.retrofit2:retrofit:2.1.0'
-    compile 'com.squareup.retrofit2:converter-gson:2.1.0'
+
+    compile "com.shopify:livedata-ktx:2.0.0"
+
+    //retrofit
     compile 'com.google.code.gson:gson:2.7'
+    compile 'com.squareup.retrofit2:adapter-rxjava:2.3.0'
+    compile 'com.squareup.retrofit2:adapter-rxjava2:2.4.0'
+    compile 'com.squareup.retrofit2:retrofit:2.4.0'
+    compile 'com.squareup.retrofit2:converter-gson:2.1.0'
+
+    //RX2
+    compile 'io.reactivex.rxjava2:rxjava:2.0.2'
+    compile 'io.reactivex.rxjava2:rxandroid:2.0.1'
+    //ViewModel
+    compile 'android.arch.lifecycle:extensions:1.1.0'
+    compile 'android.arch.lifecycle:viewmodel:1.1.0'
 }
