@@ -1,5 +1,6 @@
 package com.example.abhim.popularmovies.Activities;
 
+import android.app.Activity;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 
@@ -13,9 +14,6 @@ public class MainActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        if (savedInstanceState == null)
-            getSupportFragmentManager().beginTransaction()
-                    .add(R.id.main_activity_id, new MoviesFragment()).commit();
     }
 
 }
